Backend/requirements.txt
fastapi
uvicorn
bcrypt
python-dotenv
sqlalchemy
psycopg2-binary
python-jose[cryptography]
pyjwt
asyncpg
 fastapi
numpy 
pandas 
statsmodels 
pytest 
uvicorn